requirements.txt
numpy
scipy
pandas
loguru
deepspeed
torch==2.5.1
transformers
flash-attn
PyWavelets
matplotlib
tqdm
sktime
vllm==0.6.6.post1
ragas==0.1.9
langchain==0.2.15
langchain-chroma==0.1.1
langchain-community==0.2.15
langchain-core==0.2.37
langchain-experimental==0.0.64
langchain-openai==0.1.23
langchain-text-splitters==0.2.0
langchainhub==0.1.16
